requirements.txt
email-validator>=2.3.0
flask-cors>=6.0.1
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
pillow>=11.3.0
sqlalchemy>=2.0.43
python-dotenv>=1.0.0
